# pkg/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Message:
    properties:
      message:
        description: A standard message.
        type: string
    type: object
...
  models.UserGeneralSettings:
    properties:
      default_project_id:
        type: integer
      discoverable_by_email:
        type: boolean
      discoverable_by_name:
        type: boolean
      email_reminders_enabled:
        type: boolean
      extra_settings_links:
        additionalProperties: {}
        description: Server/OpenID-provided; populated on read, ignored on write.
        type: object
      frontend_settings: {}
      language:
        type: string
      name:
        type: string
      overdue_tasks_reminders_enabled:
        type: boolean
      overdue_tasks_reminders_time:
        type: string
      timezone:
        type: string
      week_start:
        maximum: 6
        minimum: 0
        type: integer
    type: object
...
  v1.UserAvatarProvider:
    properties:
      avatar_provider:
        description: The avatar provider. Valid types are `gravatar` (uses the user
          email), `upload`, `initials`, `marble` (generates a random avatar for each
          user), `ldap` (synced from LDAP server), `openid` (synced from OpenID provider),
          `default`.
        type: string
    type: object
...
  web.HTTPError:
    properties:
      code:
        type: integer
      i18n_params:
        additionalProperties:
          type: string
        description: I18nParams carries Message's dynamic values, keyed by the client's
          translation placeholder names, so clients can localise the error.
        type: object
      message:
        type: string
    type: object
info:
  contact:
    name: Norna
    url: https://github.com/zurdi15/norna
  description: |-
    # Pagination
    Every endpoint capable of pagination will return two headers:
    * `x-pagination-total-pages`: The total number of available pages for this request
    * `x-pagination-result-count`: The number of items returned for this request.
    # Permissions
    All endpoints which return a single item (project, task, etc.) - no array - will also return a `x-max-permission` header with the max permission the user has on this item as an int where `0` is `Read Only`, `1` is `Read & Write` and `2` is `Admin`.
    This can be used to show or hide ui elements based on the permissions the user has.
    # Errors
    All errors have an error code and a human-readable error message in addition to the http status code. You should always check for the status code in the response, not only the http status code.
    Due to limitations in the swagger library we're using for this document, only one error per http status code is documented here. The `code` field in an error response identifies the exact error.
    # Authorization
    **JWT-Auth:** Main authorization method, used for most of the requests. Needs `Authorization: Bearer <jwt-token>`-header to authenticate successfully.

...

    **BasicAuth:** Only used when requesting tasks via CalDAV.
    <!-- ReDoc-Inject: <security-definitions> -->
  license:
    name: AGPL-3.0-or-later
...
  title: Norna API
paths:
...
  /user/settings/avatar:
    get:
      consumes:
      - application/json
      description: Returns the current user's avatar setting.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserAvatarProvider'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Return user avatar setting
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Changes the user avatar. Valid types are gravatar (uses the user
        email), upload, initials, marble, ldap (synced from LDAP server), openid (synced
        from OpenID provider), default.
      parameters:
      - description: The user's avatar setting
        in: body
        name: avatar
        required: true
        schema:
          $ref: '#/definitions/v1.UserAvatarProvider'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Set the user's avatar
      tags:
      - user
...
  /user/settings/general:
    post:
      consumes:
      - application/json
      parameters:
      - description: The updated user settings
        in: body
        name: avatar
        required: true
        schema:
          $ref: '#/definitions/models.UserGeneralSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Change general user settings of the current user.
      tags:
      - user
...
  /user/timezones:
    get:
      consumes:
      - application/json
      description: Because available time zones depend on the system Norna is running
        on, this endpoint returns a project of all valid time zones this particular
        Norna instance can handle. The project of time zones is not sorted, you should
        sort it on the client.
      produces:
      - application/json
      responses:
        "200":
          description: All available time zones.
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Get all available time zones on this Norna instance
      tags:
      - user
...
securityDefinitions:
  BasicAuth:
    type: basic
  JWTKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"